+++ scripts/cardlist.js
@@ -1,17 +1,18 @@
-// import Card from './card.js'
-// export default class CardList {
-//     constructor(cardsContainer, cardsArr){
-//         this.cardsContainer = cardsContainer;
-//         this.cardsArr = cardsArr;
-//         this.render();
-//     }
+import Card from './card.js';
 
-//     addCard(crdName, crdLink){
-//         const { cardElement } = new Card(crdLink, crdName);
-//         this.cardsContainer.appendChild(cardElement);
-//     }
+export default class CardList {
+    constructor(cardsContainer, cardsArr){
+        this.cardsContainer = cardsContainer;
+        this.cardsArr = cardsArr;
+        this.render();
+    }
 
-//     render () {
-//         this.cardsArr.forEach((elem) => {this.addCard(elem.link, elem.name)})
-//     }
-// }
+    addCard(crdName, crdLink){
+        const { cardElement } = new Card(crdLink, crdName);
+        this.cardsContainer.appendChild(cardElement);
+    }
+
+    render () {
+        this.cardsArr.forEach((elem) => {this.addCard(elem.link, elem.name)})
+    }
+}
